1.0-night.rev48
Added in interface for listing out and uninstall linux icons by name.
Test it works. Did not test uninstalling linux icons by startsWith yet.

+++ modules/workers/ICGenWorker.js
@@ -1408,7 +1408,11 @@ function uninstallLinuxIconByName(aNameArr, onlyCheckTheseSizes, startsWith) {
 				hicolorEntries = hicolorIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
 			} catch(iterex) {
 				console.error('error on hicolorIterator:', iterex);
-				throw iterex;
+				// throw iterex;
+				return [{
+					status: 'fail',
+					reason: 'failed iteration on: ' + dirpathHicolor
+				}];
 			} finally {
 				hicolorIterator.close();
 			}
@@ -1424,7 +1428,7 @@ function uninstallLinuxIconByName(aNameArr, onlyCheckTheseSizes, startsWith) {
 			if (!startsWith) {
 				for (var j=0; j<aNameArr.length; j++) {
 					try {
-						OS.File.remove(OS.Path.join(dirpathSizeArr[i], aNameArr[j] + '.png'));
+						OS.File.remove(OS.Path.join(dirpathSizeArr[i], aNameArr[j] + '.png'), {ignoreAbsent:false});
 						console.log('deleted:', OS.Path.join(dirpathSizeArr[i], aNameArr[j] + '.png')); // .remove throws when file DNE, so it will not get to this line if file DNE // :note: I LIKE HOW OS.File.remove throws when file does not exist
 					} catch (filex) {
 						// i thin kthis will catch if file does not exist
@@ -1440,15 +1444,21 @@ function uninstallLinuxIconByName(aNameArr, onlyCheckTheseSizes, startsWith) {
 					dirsizeEntries = dirsizeIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
 				} catch(iterex) {
 					console.error('error on dirsizeIterator:', iterex);
-					throw iterex;
+					// throw iterex;
+					return [{
+						status: 'fail',
+						reason: 'failed iteration on: ' + dirpathHicolor
+					}];
 				} finally {
 					dirsizeIterator.close();
 				}
-				for (var j=0; i<dirsizeEntries.length; j++) {
-					for (var k=0; k<aNameArr.length; k++) {
-						if (dirsizeEntries[j].name.toLowerCase().indexOf(aNameArr[k]) == 0) {
-							OS.File.remove(dirsizeEntries[j].path);
-							console.log('deleted:', dirsizeEntries[j].path);
+				if (dirsizeEntries.length) {
+					for (var j=0; j<dirsizeEntries.length; j++) {
+						for (var k=0; k<aNameArr.length; k++) {
+							if (dirsizeEntries[j].name.toLowerCase().indexOf(aNameArr[k]) == 0) {
+								OS.File.remove(dirsizeEntries[j].path);
+								console.log('deleted:', dirsizeEntries[j].path);
+							}
 						}
 					}
 				}
@@ -1461,89 +1471,107 @@ function uninstallLinuxIconByName(aNameArr, onlyCheckTheseSizes, startsWith) {
 		console.error('platform not supported');
 	}
 	
+	return [{status:'ok'}];
+	
 }
 
 function listAllLocalIcons() {
 	// this is specific for Icon Container Generator addon
 	
-	var rezObj = {}; // key is name, and value is array of dir names it was found in
-	
-	var dirpathHicolor = OS.Path.join(
-		OS.Constants.Path.homeDir,
-		'.local',
-		'share',
-		'icons',
-		'hicolor'
-	);
+	if (core.os.toolkit.indexOf('gtk') == 0) {
+		var rezObj = {}; // key is name, and value is array of dir names it was found in
+		
+		var dirpathHicolor = OS.Path.join(
+			OS.Constants.Path.homeDir,
+			'.local',
+			'share',
+			'icons',
+			'hicolor'
+		);
 
-	var dirpathSizeArr = [];
-	var dirnameSizeArr = [];
-	
-	var hicolorIterator;
-	var hicolorEntries;
-	try {
-		hicolorIterator = new OS.File.DirectoryIterator(dirpathHicolor);
-		hicolorEntries = hicolorIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
-	} catch(iterex) {
-		console.error('error on hicolorIterator:', iterex);
-		throw iterex;
-	} finally {
-		hicolorIterator.close();
-	}
-	for (var i=0; i<hicolorEntries.length; i++) {
-		if (hicolorEntries[i].isDir) {
-			dirpathSizeArr.push(OS.Path.join(hicolorEntries[i].path, 'apps'));
-			dirnameSizeArr.push(hicolorEntries[i].name); // dirnameSizeArr[i] == i of dirpathSizeArr // link112005145
-		}
-	}
+		var dirpathSizeArr = [];
+		var dirnameSizeArr = [];
 		
-	
-	for (var i=0; i<dirpathSizeArr.length; i++) {
-		var dirsizeIterator;
-		var dirsizeEntries;
-		
-		var cDirSizeNameFriendly = dirnameSizeArr[i];
-		var numMatch;
-		if (cDirSizeNameFriendly.toLowerCase() == 'scalable') {
-			cDirSizeNameFriendly = 'SVG';
-		} else if (numMatch = cDirSizeNameFriendly.exec(/\d+/)) {
-			cDirSizeNameFriendly = numMatch[0];
-		}
-		
+		var hicolorIterator;
+		var hicolorEntries;
 		try {
-			dirsizeIterator = new OS.File.DirectoryIterator(dirpathSizeArr[i]);
-			dirsizeEntries = dirsizeIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
+			hicolorIterator = new OS.File.DirectoryIterator(dirpathHicolor);
+			hicolorEntries = hicolorIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
 		} catch(iterex) {
-			console.error('error on dirsizeIterator:', iterex);
-			throw iterex;
+			console.error('error on hicolorIterator:', iterex);
+			// throw iterex;
+			return [{
+				status: 'fail',
+				reason: 'failed iteration on: ' + dirpathHicolor
+			}];
 		} finally {
-			dirsizeIterator.close();
+			hicolorIterator.close();
 		}
-		for (var j=0; i<dirsizeEntries.length; j++) {
-			var nameNoExt = dirsizeEntries[j].name.substr(0, dirsizeEntries[j].name.lastIndexOf('.'));
-			if (!(nameNoExt in rezObj)) {
-				rezObj[nameNoExt] = [];
-				
-					sizes: [], // friendly, like #x# is turned to # and scalable to svg
-					sizeDirName: [], // list of dir names
-					nameWithExt: []
-				};
-			}
-			
-			rezObj[nameNoExt].push({
-				sizeFriendly: cDirSizeNameFriendly
-				sizeDirName: dirnameSizeArr[i], // dirnameSizeArr[i] == i of dirpathSizeArr // link112005145
-				nameWithExt: dirsizeEntries[j].name // key with the extension of .png or .svg, i do this cuz i want to know if some .png/.svg are caps or mixed casing
-			});
-			
-			if (dirsizeEntries[j].name.toLowerCase().indexOf(aNameArr[k]) == 0) {
-				OS.File.remove(dirsizeEntries[j].path);
-				console.log('deleted:', dirsizeEntries[j].path);
+		for (var i=0; i<hicolorEntries.length; i++) {
+			if (hicolorEntries[i].isDir) {
+				dirpathSizeArr.push(OS.Path.join(hicolorEntries[i].path, 'apps'));
+				dirnameSizeArr.push(hicolorEntries[i].name); // dirnameSizeArr[i] == i of dirpathSizeArr // link112005145
 			}
 		}
+			
+		
+		for (var i=0; i<dirpathSizeArr.length; i++) {
+			var dirsizeIterator;
+			var dirsizeEntries;
+			
+			var cDirSizeNameFriendly = dirnameSizeArr[i];
+			var numMatch;
+			if (cDirSizeNameFriendly.toLowerCase() == 'scalable') {
+				cDirSizeNameFriendly = 'SVG';
+			} else if (numMatch = /\d+/.exec(cDirSizeNameFriendly)) {
+				cDirSizeNameFriendly = parseInt(numMatch[0]);
+			}
+			
+			try {
+				dirsizeIterator = new OS.File.DirectoryIterator(dirpathSizeArr[i]);
+				dirsizeEntries = dirsizeIterator.nextBatch();  // this will throw if path at str doesnt exist, this only happens on pathToDir though, as the rest is on stuff thats found link10000002551 i got this: `message:"Error: Win error 2 during operation DirectoryIterator.prototype.next on file C:\Users\Vayeate\AppData\Roaming\Mozilla\Firefox\profilist_data\launcher_exes (The system cannot find the file specified.)`
+			} catch(iterex) {
+				console.error('error on dirsizeIterator:', iterex);
+				// throw iterex;
+				return [{
+					status: 'fail',
+					reason: 'failed iteration on: ' + dirpathSizeArr[i]
+				}];
+			} finally {
+				dirsizeIterator.close();
+			}
+			if (dirsizeEntries.length) {
+				for (var j=0; j<dirsizeEntries.length; j++) {
+					var nameNoExt = dirsizeEntries[j].name.substr(0, dirsizeEntries[j].name.lastIndexOf('.'));
+					if (!(nameNoExt in rezObj)) {
+						rezObj[nameNoExt] = [];
+					}
+					
+					rezObj[nameNoExt].push({
+						sizeFriendly: cDirSizeNameFriendly,
+						sizeDirName: dirnameSizeArr[i], // dirnameSizeArr[i] == i of dirpathSizeArr // link112005145
+						nameWithExt: dirsizeEntries[j].name, // key with the extension of .png or .svg, i do this cuz i want to know if some .png/.svg are caps or mixed casing
+						fileUri: OS.Path.toFileURI(dirsizeEntries[j].path)
+					});
+				}
+			}
+		}
+		
+		return [{
+			iconsList: rezObj,
+			status: 'ok'
+		}];
+	} else if (core.os.toolkit.indexOf('qt') == 0) {
+		return [{
+			reason: 'QT platform not yet supported',
+			status: 'fail'
+		}];		
+	} else {
+		return [{
+			reason: 'Your operating system is not supported',
+			status: 'fail'
+		}];		
 	}
-	
-	return rezObj;
 }
 // End - Addon Functionality
